+++ e2e/pages/base.py
